requirement.txt
aiofiles==22.1.0
aiogoogle==4.3.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
black==22.8.0
cachetools==5.2.0
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.1
google-api-core==2.10.0
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
greenlet==1.1.3
httplib2==0.20.4
identify==2.5.4
idna==3.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
nodeenv==1.7.0
pathspec==0.10.1
platformdirs==2.5.2
pre-commit==2.20.0
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.40
toml==0.10.2
tonyg-rfc3339==0.1
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.4
Werkzeug==2.2.2
yarl==1.8.1
